requirements.txt
openai>=1.3.7
pinecone>=3.0.0
python-dotenv
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic[email]==2.4.2